+++ app/controllers/authenticate/sessions_controller.rb
@@ -3,6 +3,7 @@
 #
 class Authenticate::SessionsController < Authenticate::AuthenticateController
   before_action :redirect_signed_in_users, only: [:new]
+  skip_before_action :require_login, only: [:create, :new, :destroy], raise: false
   skip_before_action :require_authentication, only: [:create, :new, :destroy], raise: false
 
   #
@@ -38,7 +39,7 @@ class Authenticate::SessionsController < Authenticate::AuthenticateController
   private
 
   def redirect_signed_in_users
-    redirect_to url_for_signed_in_users if authenticated?
+    redirect_to url_for_signed_in_users if logged_in?
   end
 
   def url_after_create