+++ spec/dummy/app/controllers/application_controller.rb
@@ -1,6 +1,7 @@
 class ApplicationController < ActionController::Base
   include Authenticate::Controller
-  before_action :require_authentication
+  # before_action :require_authentication
+  before_action :require_login
 
   # Prevent CSRF attacks by raising an exception.
   # For APIs, you may want to use :null_session instead.