+++ spec/dummy/config/initializers/authenticate.rb
@@ -5,4 +5,5 @@ Authenticate.configure do |config|
   config.bad_login_lockout_period = 10.minutes
   config.reset_password_within = 5.minutes
   config.password_length = 8..128
+  config.debug = true
 end